updated lambda methods that directly called other methods for sonic-lint java:S1612
specific to Register.java, the action listeners for password, username, and email were
referring to empty methods, so just added a lambda method with a comment.  For the register
button, I left it as a separate method but it doesn't use the event body so I just removed
the event parameter from the function signature, and lint is happy because now it isn't
simply passing "event" along from the lambda method to a proper method call.