updated lambda methods that directly called other methods for sonic-lint java:S1612
Firstly, a lot of these popped up when I fixed the java:S1604 warnings, so I wanted
to continue until I got them more appropriately dealt with.

Largely this sonic-lint warning was about having lambda methods that simply passed
their argument along to another method with the same signature.  For the majority
of these fixes, I was able to migrate the method body from the called method into
a lambda function that was inline with the addListener call.

For a few that weren't more conveniently defined in code, either for no method body
at all or a long one, I made (hopefully) judicious choices to try to keep from having
the inline code become uselessly long, and also conserve the original intent by
commenting when nothing was done.