Converting anonymous inner classes with one method into lambdas (#40)
* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted getStuff from Runnable instance to lambda method

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted single-method ActionListener methods to lambdas that call the assigned methods

* converted startExpander's thread to a lambda function to clear sonic-lint java:1604

* converted getStuff's runnable to a lambda function to clear sonic-lint java:1604

* updated lambda methods that directly called other methods for sonic-lint java:S1612

specific to Register.java, the action listeners for password, username, and email were
referring to empty methods, so just added a lambda method with a comment.  For the register
button, I left it as a separate method but it doesn't use the event body so I just removed
the event parameter from the function signature, and lint is happy because now it isn't
simply passing "event" along from the lambda method to a proper method call.

* updated lambda methods that directly called other methods for sonic-lint java:S1612

specific to SpellList1.java, the action listeners for several buttons were all loading
consecutive spells with nearly identical code, so I consolidated those into a single
method, and then had the lambda methods reference it with appropriate start indices.

* updated lambda methods that directly called other methods for sonic-lint java:S1612

Firstly, a lot of these popped up when I fixed the java:S1604 warnings, so I wanted
to continue until I got them more appropriately dealt with.

Largely this sonic-lint warning was about having lambda methods that simply passed
their argument along to another method with the same signature.  For the majority
of these fixes, I was able to migrate the method body from the called method into
a lambda function that was inline with the addListener call.

For a few that weren't more conveniently defined in code, either for no method body
at all or a long one, I made (hopefully) judicious choices to try to keep from having
the inline code become uselessly long, and also conserve the original intent by
commenting when nothing was done.